Clear up warnings
* ISO C99 doesn’t support unnamed structs/unions
    -Modified the unamed unions in oc_rep_t and oc_endpoint_t,
     and updated all affected modules.
* ISO C forbids conversion of object pointer to function
  pointer type
    -Factored out the response and discovery handlers into a
     client_handler union which is stored in oc_client_cb_t.
     This way each type of handler can be explicitly stored
     and invoked as such instead of storing them as a void *
     and doing a conversion.
* ISO C forbids an empty translation unit
* unused-parameter
* missing-field-initializers
* sign-compare
* pointer-sign
* unused-but-set-variable
* strict-prototypes

+++ api/oc_blockwise.c
@@ -79,7 +79,7 @@ oc_blockwise_alloc_response_buffer(const char *href, int href_len,
     int i = COAP_ETAG_LEN;
     uint32_t r = oc_random_value();
     while (i > 0) {
-      memcpy(buffer->etag, &r, MIN(sizeof(r), i));
+      memcpy(buffer->etag, &r, MIN((int)sizeof(r), i));
       i -= sizeof(r);
       r = oc_random_value();
     }
